refactor: extract WorkoutItem component from erg-dashboard.jsx
Moves ~220-line component (+ workoutAccent helper) to
web/src/components/WorkoutItem.jsx. Used in calendar, program,
and overview views.

Build and 68 tests pass.
